// src/payload-types.ts
export interface Site {
  id: number;
  contactEmail?: string | null;
  social?:
    | {
        label: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  footerNote?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}